+++ renderer/init/Bootstrap.js
@@ -11,7 +11,7 @@ import { blobManager } from '@core/BlobManager.js';
 import { initSharedModals } from '@core/SharedModal.js';
 import { updateManager } from '@core/UpdateManager.js';
 import { syntaxHighlighter } from '@core/syntaxHighlighter/SyntaxHighlighter.js';
-import { setCodeHighlighter } from '@common/MarkdownParser.js';
+import { setCodeHighlighter, setHtmlCodeLineSplitter } from '@common/MarkdownParser.js';
 
 import { registerGlobalEvents } from './InitEvents.js';
 import { registerPresets } from './InitPresets.js';
@@ -42,6 +42,9 @@ export async function bootstrap() {
   syntaxHighlighter.warmpUp();
   setCodeHighlighter(({ langId, styleId, text }) =>
     syntaxHighlighter.highlightTextAsHTML({ langId, styleId, text })
+  );
+  setHtmlCodeLineSplitter((htmlCodeStr) => 
+    syntaxHighlighter.splitHighlightedHtmlIntoLines(htmlCodeStr)
   );
 
   viewManager.init(document.getElementById('app'));